% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_variance.R
\name{my_variance}
\alias{my_variance}
\title{Sample Variance}
\usage{
my_variance(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the sample variance
}
\description{
The function takes a vector and calculates the
sample variance
}
\examples{
x  <- 1:10
my_variance(x)
}
